requirements.txt
torch 
torchvision 
torchaudio 
torchmetrics>=1.0.1
numpy==1.26.3
pandas>=1.3.4
matplotlib>=3.4.3
tqdm>=4.62.3
scikit-learn>=0.24.2
scikit-image>=0.18.3
pillow >=8.4.0
opencv-python
